Changed DMA word num to u32 and fixed formatting

+++ libpsx/src/dma/mod.rs
@@ -8,7 +8,7 @@ rw_register!(GpuDmaControl, 0x1F80_10A8);
 
 pub enum DmaBlocks {
     // TODO: this should be u32 since 0x10000 is valid and gets mapped to 0u16
-    Words(u16),
-    Blocks { words: u16, blocks: u16 },
+    Words(u32),
+    Blocks { words: u32, blocks: u32 },
     LinkedList,
 }
